+++ SCF/BibliotecaSCF/Clases/ItemNotaDeCredito.cs
@@ -10,7 +10,7 @@ namespace BibliotecaSCF.Clases
     public class ItemNotaDeCredito
     {
         public virtual int Codigo { get; set; }
-        public virtual int Cantidad { get; set; }
+        public virtual double Cantidad { get; set; }
         public virtual ItemEntrega ItemEntrega { get; set; }
 
         private static DataTable Tabla
