+++ SCF/BibliotecaSCF/Clases/ItemNotaDePedido.cs
@@ -9,7 +9,7 @@ namespace BibliotecaSCF.Clases
     public class ItemNotaDePedido
     {
         public virtual int Codigo { get; set; }
-        public virtual int CantidadPedida { get; set; }
+        public virtual double CantidadPedida { get; set; }
         public virtual DateTime FechaEntrega { get; set; }
         public virtual double Precio { get; set; }
         public virtual int Posicion { get; set; }
